requirements.txt
numpy<2.0.0
streamlit==1.32.0
pandas==2.2.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
plotly==5.20.0